+++ src/test/java/org/opensearch/knn/plugin/action/RestKNNStatsHandlerIT.java
@@ -7,8 +7,6 @@ package org.opensearch.knn.plugin.action;
 
 import lombok.SneakyThrows;
 import org.apache.hc.core5.http.io.entity.EntityUtils;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.junit.Before;
 import org.junit.rules.DisableOnDebug;
 import org.opensearch.client.Request;
@@ -57,7 +55,6 @@ import static org.opensearch.knn.common.KNNConstants.NAME;
  */
 public class RestKNNStatsHandlerIT extends KNNRestTestCase {
 
-    private static final Logger logger = LogManager.getLogger(RestKNNStatsHandlerIT.class);
     private static final String TRAINING_INDEX = "training-index";
     private static final String TRAINING_FIELD = "training-field";
     private static final String TEST_MODEL_ID = "model-id";